+++ src/ignitus-Routes/publicRoutes.js
@@ -14,7 +14,7 @@ import LoginProfessor from "../ignitus-ProfessorLogin";
 import SignupProfessor from "../ignitus-ProfessorSignUp";
 import LoginStudent from "../ignitus-StudentLogin";
 import SignupStudent from "../ignitus-StudentSignUp";
-import { CommonLoginUI, CommonSignUpUI } from "../ignitus-Common-Auth-Student-Prof/Components/createSignInPage";
+import { CommonLoginUI, CommonSignUpUI } from "../ignitus-Common-Auth-Student-Prof/Components/commonLoginRegister";
 
 import loader from "../ignitus-Assets/Images/loader2.gif";
 import Loadable from "react-loadable";
